change: refactor types into the /types directory
